generate an xml file with the filename and title
The idea is to have an xml file which stores the name of the xml
file and the title of each post. This was difficult, because `msxsl`
has problems with `document()`; therefore, I downloaded libxslt (with
all the dependencies). It is used as such:
`xsltproc -o posts_with_titles.xml make_posts_with_titles.xsl posts.xml`

This file will be loaded by "index.xml", and a list created with all the
posts. When the user hovers over a post's title, additional information
will be fetched via Ajax - such as date, author, tags.

These links proved helpfull:
http://stackoverflow.com/a/6815721/447661
http://stackoverflow.com/questions/4824843/adding-attribute-to-the-node
http://www.dpawson.co.uk/xsl/sect2/identity.html (the last reply)

The difference between `copy` and `copy-of` is also very important;
it is discussed in "Mastering XSLT", pages 132-139.