fix js initialization bug on Chrome
`window.onload` is supposed to be run when the page load is complete;
however, this does not seem to be the case with Chrome; calling the
initialization function in the header does not work (but it does in
Firefox). Fix this by calling the initialization function just before
closing the `body` tag.